Add support for jujutsu workspaces.
Jujutsu workspaces are basically the equivalent of git worktrees. This
change adds support for those by checking if `repo` is a directory, and
if not, reading it to find the upstream repository path.

Testing Done:
Posted a change out of a jujutsu workspace.

Reviewed at https://reviews.reviewboard.org/r/14564/