Add Git binary file patching.
This change adds support to the `GitPatcher` for applying binary files
as part of changes. For git, this requires that we ask `git apply` to
exclude those particular files while applying, because otherwise it will
error out because there's no binary file data in the patch itself. We
then use the base patcher infrastructure for iterating through the
files, but provide overrides that can register adds, moves, and deletes
with git.

Testing Done:
- Ran unit tests.
- Used `rbt patch` with a bunch of changes that did various file
  operations on binary files.

Reviewed at https://reviews.reviewboard.org/r/14714/