Add SVN binary file patching.
This change adds support to the `SVNPatcher` for applying binary files
as part of changes. Things have been reorganized a little bit in order
to make it easier to deal with both empty files and binary files, which
appear similarly in the diff.

Additionally, several of our tests had index lines that did not start
with a `/`, which was inconsistent with real-world data. I've fixed
these up to match.

Testing Done:
- Ran unit tests.
- Used `rbt patch` with a bunch of changes that did various file
  operations on binary files.

Reviewed at https://reviews.reviewboard.org/r/14717/