Switch to getting appdirs from pypi.
When we first added `appdirs` to RBTools, we vendored the file into our
tree because they were not providing builds on PyPI. This changed a long
time ago, so there's no good reason for us to continue to ship our own,
outdated copy.

Testing Done:
Ran unit tests.
Fix patch tests to not use deprecated names.

This fixes up the patch tests to use `full_name` instead of the
deprecated `fullname` attribute for authors. This also updates one test
that passes in an invalid value for `prefix_level` to properly suppress
the expected deprecation warning.

Testing Done:
Ran unit tests.
Add my linter configuration
Fix parsing of `git diff-tree` output with copied/renamed files.

When a git change includes copied files and we have exclude patterns set
(either via the config file or the command-line), we were not parsing
the `git diff-tree` output correctly.

This change redoes the parsing to use the `-z` flag, which splits
records with NUL characters instead of tabs (making us more resilient
with filenames). According to the diff-tree docs, the resulting records
may include either one or two files based on the change marker. If we
have a "C" (copy) or "R" (rename) change, we use the second filename as
the test for whether we should exclude the result from the diff.
Otherwise, we parse out a single filename and test that.

A new unit test has been added which failed with the initial
implementation and succeeds with the new one.

Testing Done:
- Verified that diffs were properly created with the reported situation.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14577/